Exception handling

+++ src/main/java/com/caue/democommerce/services/ProductService.java
@@ -3,7 +3,10 @@ package com.caue.democommerce.services;
 import com.caue.democommerce.dto.ProductDTO;
 import com.caue.democommerce.entities.Product;
 import com.caue.democommerce.repositories.ProductRepository;
+import com.caue.democommerce.services.exceptions.DatabaseException;
+import com.caue.democommerce.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -19,7 +22,7 @@ public class ProductService {
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
 
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Product not found!"));
         return new ProductDTO(product);
     }
 
@@ -41,6 +44,10 @@ public class ProductService {
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
 
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Product not found to be updated.");
+        }
+
         Product entity = repository.getReferenceById(id);
         copyDtoToEntity(entity, dto);
         entity = repository.save(entity);
@@ -48,8 +55,20 @@ public class ProductService {
     }
 
     @Transactional
-    public void delete(Long id){
-        repository.deleteById(id);
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Product not found.");
+        }
+
+        try {
+            repository.deleteById(id);
+            repository.flush();
+
+        } catch (DataIntegrityViolationException e) {
+
+            throw new DatabaseException("Database integrity violated.");
+        }
+
     }
 
     private void copyDtoToEntity(Product entity, ProductDTO dto) {
